Allow UDP traceroute in common-linux role.
This fix is in response to RT ticket #14 "Fix firewall in Tukwila", where Bart
discovered that UDP traceroute does not work to VMs at Tukwila. These VMs were
configured with the common-linux role.

https://issues.hamwan.net/rt/Ticket/Display.html?id=14

Unixy traceroute uses UDP starting on port 33434 and incrementing by 1 for each
probe. This means we must open UDP between 33434 and a maximum of
33434 + (max-ttl * probes-per-hop - 1). Accounting for a liberal max-ttl of 64
and probes-per-hop of 3 gives us a maximum of port 33625.

http://superuser.com/a/807465/104150

This patch opens UDP port 33434-33625.