Expand roles/users tests to cover changing of group names
Introduced a new Vagrant target "test.group.change" to test the scenario of
an admin user being in an outdated group (eg: hamwan), and being automatically
moved to a new group (eg: hamadmin).  The scenario injects multiple network
failures, which exposes an unsafe sequencing of tasks.  The admin user is locked
out of network control, and the test fails.

In order to pull this off, I had to stop using production keys, and there's now
a locally stored test key pair.  Don't use it in prod or for anything else!
The private key is stored in plaintext and distributed widely.

The Ansible admin user changed from "vagrant" to "eo".  Vagrant user remains for
true back-door access to the test target VMs.