Fully decouple all role variables for testing & update production users
The tests relied on the role's default variables, but now the test variables
are fully isolated.  This improves test stability.  For example, adding users
to the role variables caused tests to break for lack of test keys for said new
users.

This change also updates the live production user list with the results of a
recent audit.  This is now the authoritative list of users to distribute to
prod.

Lastly, some lookup() plugin names were incorrect and are now fixed.