Restructure handling of site and host specific settings. The good news is that now all of the site specific information is now in group_vars/ or host_vars/ and not in the role.  Host specific settings are now fairly trivial to implement.
This was a fairly major change and required a couple of subtle coding solutions
to handle things properly.
The first is to reference the list of settings in a group_vars file to allow it to
be overriden (see os_routeros.yml and routeros_common/vars/main.yml).
Another issue was the handling of a setting that used the word 'no', which
Ansible/Jinga insisted in coverting to a bool False.  I solved this by changing
the desired setting to '=no' and updated the rest of that stanza accordingly.

Other changes:
Updated host overrides for Triangle machines per the new structure.
Cosmetic change to the diff printout so you can see the ACTION in all cases.
Cosmetic change to some quoting (" -> ')