Ansible updates to support common routeros settings (#22)
This is the initial set of changes to support RouterOS under Ansible mangement for HamWAN.
A subsequent set of changes will provide molecule/vagrant testing, similar to the existing users role.

* Additional refinement of the new role routeros_common.

* Add missing files and lint updates.

* Add emacs/mg editor backfiles to .gitignore

* Added stanza for /ip dns servers setting

* Added configuration of dns allow-remote-requests.

* Add variable relevant to all routeros hosts/connections.

* Improve handling of header lines and reduce unneeded debugging output.

* Improve handling of values that are OK if missing.
This was necessary to be able to avoid showing a change was needed when
in fact it wasn't (dhcp_dns_servers, dhcp_ntp_servers).  We were invoking the
change command which was a no-op in these cases, but it was messy and
with a little more logic.

* Add missing_ok variable to all settings.

* Make snmp community reference more specific.

* Adding new files, forgotten in earlier commits.
get_facts.yml is only used for getting routeros version for now.
psdr_routeros.yml is a convenience playbook that only processes routeros plays.

* Use export instead of print for more reliable parsing of output.

* Add ability to override standard settings on a per-host basis.
Null_settings.yml is the do nothing option (no settings in any list).
The S?.Triangle files show how standard settings can be overridden where needed.
Also added .log files to .gitignore.

* Update pattern for log files.

* Fix DNS settings and add vars file for Lookout.Triangle.

* Add tags to psdr.yml and remove psdr_routeros.yml.
Tags are a better way to exercise only a portion of the playbook.

* Improve file comments and add TODOs to make settings.yml site independent.

* First pass of code review driven changes.
The principle change is to eliminate code duplication by having check_and_set_value.yml
call get_facts.yml.  I needed some minor changes to the later to enable this.
Added a README for the routeros_common role and improved some comments.
Fixed some errors in the Triangle override files but the override mechanism
is likely to change in the next pass. Cosmetic changes to psdr.yml.
Removed unnecessary use of reserved variable in vars/main.yml.

* Restructure handling of site and host specific settings.
The good news is that now all of the site specific information is now in
group_vars/ or host_vars/ and not in the role.  Host specific settings are now
fairly trivial to implement.

This was a fairly major change and required a couple of subtle coding solutions
to handle things properly.
The first is to reference the list of settings in a group_vars file to allow it to
be overriden (see os_routeros.yml and routeros_common/vars/main.yml).
Another issue was the handling of a setting that used the word 'no', which
Ansible/Jinga insisted in coverting to a bool False.  I solved this by changing
the desired setting to '=no' and updated the rest of that stanza accordingly.

Other changes:
Updated host overrides for Triangle machines per the new structure.
Cosmetic change to the diff printout so you can see the ACTION in all cases.
Cosmetic change to some quoting (" -> ')

* Updating comments.  No material changes.

* Updated README with better description.

* Update existing ansible testing.
We moved Vagrantfile to roles/users since it only applies to that testing.
An additional Vagrantfile will be needed for the routeros_common role.
Newer versions of molecule (since June 2023) changed the name of the delegated driver
to 'default', so we have made that change here.  Also, vagrant has started to enforce
a default range for network addresses, so we have changed the addresses used in our instances.

* Update README to add testing instructions for Debian based systems.

* Add brief comment on testing strategy.  This will be updated when
I have more experience and have written the testing for routeros_common.

* Add invocation of python virtual environment at appropriate point in setup.

* Move linux addresses to 192.168.58.0/24 to differentiate from routeros network.

* Setup first version of vagrant configuration for RouterOS.
The next commit will remove unnecessary configuraiton and make it more
like our standard configuration, but this version works.
Modeled after information at https://github.com/cheretbe/packer-routeros

* Improve completeness of setup instructions for Debian.