Update users role (#25)
* Don't try to update users on anycast "hosts".
There is no guarantee what host you will wind up connecting to if
multiple hosts provide the service.  Use the hosts specific identity
and address instead.

* Commit routeros testing support in a working state.
There is certainly some fine tuning to be done on this but the
ansible + molecule + vagrant support is wired up and working and
successfully processes testing against ROS6 and ROS7.

* Additional files needed for initial routeros testing support.

* Update reference to README for role routeros_common

* Add more description of the testing strategy for routeros_common.

* Made program portable to other users by using $HOME reference.

* Add comment about need to stay in sync with Vagrant configuration.

* This makes the choice of file explicit at the minor cost of some repetition.
It has the advantage of also being more flexible if more segregation of config
is needed.